[feat][mdsv2] Merge mdsstat to fsstat.
